enabling a method to override a) the Taggings and Tag class in general b) customize it for single contexts
in order to retreive correct models that can be customized and work with named routes. NOTE the tagger is not covered yet and one test is failing in taggable_spec due to unknown reasons. use reflect_on_all_associations to find it out. NOTE: maybe the whole approach of passing the classes sucks and we have to switch to a global constant. that would work probably just fine