Use config_accessor instead of mattr_accessor
mattr_accessor is removed in (not sure which version of)
ActiveSupport. This seems syntactically more correct than
using class variables.

I would be more interested in seeing something like a configuration
object on ActsAsTaggableOn with an initializer in which these
options could be configured. I'm not sure if there is a reason
or particular use-case for not doing that, but moving it there
would be a breaking change in the API for AATO.