Reverse check order before adding custom contexts
Reversing this order allows the process to move forward without loading
the `taggings` relationship when a context has already been defined for
the model.

This helps when a model is using a `cached_#{context.to_s.singular}_list`
column and the `taggings` are NOT being eager loaded. It can save a lot
of unexpected N+1 cases.