Fix `base_class` with Zeitwerk
Rails [Zeitwerk](https://github.com/fxn/zeitwerk) autoloader is a lot
more strict about when constants are loaded than the old autoloader.
I noticed that in
https://github.com/mbleigh/acts-as-taggable-on/pull/1079 `base_class` is
being set to the actual class in an initializer, which results in an
`unintialized constant` error when the autoloader is set to Zeitwerk.
This fixes the setting to work in Zeitwerk by allowing it to be set to
a `String` which gets constantized later, after initializers have run.

I also fixed the initializer example to call the `setup` method, which
I believe is the recommended way to do that.